requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
stripe==7.4.0
razorpay==1.3.0
reportlab==4.0.9
